install quake shareware files as part of configure stage for darkplaces quake so it will be compatible with the binary installer once it is out of experimental. use the new code from tyrquake for downloading/unpacking of the pak files - and only do so if they are not already installed. probably should split this out into a separate module, or have a single function in one of the implementation that the other one calls

+++ scriptmodules/libretrocores/tyrquake.sh
@@ -28,17 +28,18 @@ function install_tyrquake() {
 function configure_tyrquake() {
     mkdir -p "$romdir/ports/quake"
 
-    # download / unpack / install quake shareware files
-    wget "http://downloads.petrockblock.com/retropiearchives/quake106.zip" -O quake106.zip
-    unzip -o quake106.zip -d "quake106"
-    rm quake106.zip
-    pushd quake106
-    lhasa ef resource.1
-    cp -rf id1 "$romdir/ports/quake/"
-    popd
-    rm -rf quake106
-
-    chown -R $user:$user "$romdir/ports/quake"
+    if [ ! -f "$romdir/ports/quake/id1/pak0.pak" ]; then
+        # download / unpack / install quake shareware files
+        wget "http://downloads.petrockblock.com/retropiearchives/quake106.zip" -O quake106.zip
+        unzip -o quake106.zip -d "quake106"
+        rm quake106.zip
+        pushd quake106
+        lhasa ef resource.1
+        cp -rf id1 "$romdir/ports/quake/"
+        popd
+        rm -rf quake106
+        chown -R $user:$user "$romdir/ports/quake"
+    fi
 
     ensureSystemretroconfig "quake"
